IMPALA-2328 Parquet scan should use min/max stats
Parquet stores min/max stats which can be used to skip reading blocks if
they don't qualify a certain predicate.

This patch create a tuple row which consists of two tuples to store the
min/max value, and evaluate the conjuncts on it. If the [min, max] range
don't qualify the predicate, we can skip this row group.

Change-Id: I91de1f4d0fb2a982d06cd344e41901e3bf3c2cea